+++ nengo/tests/test_connection.py
@@ -615,7 +615,8 @@ def test_slicing_function(Simulator, plt, seed):
     assert np.allclose(w, y, atol=0.1)
 
 
-def test_list_indexing(Simulator, plt, seed):
+@pytest.mark.parametrize("negative_indices", (True, False))
+def test_list_indexing(Simulator, plt, seed, negative_indices):
 
     with nengo.Network(seed=seed) as model:
         u = nengo.Node([-1, 1])
@@ -623,10 +624,16 @@ def test_list_indexing(Simulator, plt, seed):
         b = nengo.Ensemble(40, dimensions=1, radius=2.2)
         c = nengo.Ensemble(80, dimensions=2, radius=1.3)
         d = nengo.Ensemble(80, dimensions=2, radius=1.3)
-        nengo.Connection(u[[0, 1]], a[[0, 0]])
-        nengo.Connection(u[[1, 1]], b[[0, 0]])
-        nengo.Connection(u[[0, 1]], c[[0, 1]])
-        nengo.Connection(u[[1, 1]], d[[0, 1]])
+        if negative_indices:
+            nengo.Connection(u[[0, 1]], a[[0, -1]])
+            nengo.Connection(u[[1, -1]], b[[0, -1]])
+            nengo.Connection(u[[0, 1]], c[[0, 1]])
+            nengo.Connection(u[[1, -1]], d[[0, 1]])
+        else:
+            nengo.Connection(u[[0, 1]], a[[0, 0]])
+            nengo.Connection(u[[1, 1]], b[[0, 0]])
+            nengo.Connection(u[[0, 1]], c[[0, 1]])
+            nengo.Connection(u[[1, 1]], d[[0, 1]])
 
         a_probe = nengo.Probe(a, synapse=0.03)
         b_probe = nengo.Probe(b, synapse=0.03)
@@ -644,16 +651,17 @@ def test_list_indexing(Simulator, plt, seed):
 
     line = plt.plot(t, a_data)
     plt.axhline(0, color=line[0].get_color())
-    assert np.allclose(a_data[t > 0.15], [0], atol=0.1)
     line = plt.plot(t, b_data)
     plt.axhline(2, color=line[0].get_color())
-    assert np.allclose(b_data[t > 0.15], [2], atol=0.1)
     line = plt.plot(t, c_data)
     plt.axhline(-1, color=line[0].get_color())
-    assert np.allclose(c_data[t > 0.15], [-1, 1], atol=0.1)
     line = plt.plot(t, d_data)
     plt.axhline(1, color=line[1].get_color())
-    assert np.allclose(d_data[t > 0.15], [1, 1], atol=0.1)
+
+    assert np.allclose(a_data[t > 0.15], [0], atol=0.15)
+    assert np.allclose(b_data[t > 0.15], [2], atol=0.15)
+    assert np.allclose(c_data[t > 0.15], [-1, 1], atol=0.15)
+    assert np.allclose(d_data[t > 0.15], [1, 1], atol=0.15)
 
 
 @pytest.mark.filterwarnings('ignore:boolean index did not match')